feat(api): 新增频道身份组和API权限相关接口支持
- 新增 ApiPermission 和 ApiPermissionsResponse DTO 类
- 新增 RolesResponse DTO 类用于封装频道身份组响应数据
- 在 BotApi 接口中增加 getRoles 和 getApiPermissions 方法
- 扩展 Role 类中的 color 字段类型从 Integer 改为 Long-为新增接口添加对应的单元测试方法 getRoles 和 getApiPermissions

+++ src/main/java/bot/dto/Role.java
@@ -12,7 +12,7 @@ public record Role(
         // 名称
         String name,
         // ARGB的HEX十六进制颜色值转换后的十进制数值
-        Integer color,
+        Long color,
         // 是否在成员列表中单独展示: 0-否, 1-是
         Integer hoist,
         // 人数
